refactor: 마이페이지 컴포넌트, 페이지 폴더 배럴 익스포트 설정(#72)
