+++ src/components/mypage/index.ts
@@ -1,0 +1,5 @@
+export { InfoBar } from "./InfoBar";
+export { MyPageTabs } from "./MyPageTab";
+export { OptionLink } from "./OptionLink";
+export { PhotoLabCard } from "./PhotoLabCard";
+export { PostCard } from "./PostCard";
