+++ src/pages/mypage/index.ts
@@ -1,0 +1,13 @@
+export { MyPage } from "./MyPage";
+
+//edit-info 폴더
+export { EditInfoPage } from "./edit-info/EditInfoPage";
+export { NickNameEditPage } from "./edit-info/NickNameEditPage";
+export { PhoneEditPage } from "./edit-info/PhoneEditPage";
+export { SocialPage } from "./edit-info/SocialPage";
+export { WithDrawPage } from "./edit-info/WithDrawPage";
+
+//tab 폴더
+export { LikedPhotoLabPage } from "./tab/LikedPhotoLabPage";
+export { LikedPostPage } from "./tab/LikedPostPage";
+export { MyPostPage } from "./tab/MyPostPage";
